CarServiceTest update exception add Dao fix

+++ service/src/main/java/cz/muni/fi/service/CarServiceImpl.java
@@ -4,7 +4,9 @@ import cz.muni.fi.dao.CarDao;
 import cz.muni.fi.dto.CarDTO;
 import cz.muni.fi.entities.Car;
 import cz.muni.fi.entities.Driver;
+import cz.muni.fi.service.exception.CustomDataAccessException;
 import org.dozer.inject.Inject;
+import org.springframework.dao.DataAccessException;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -21,37 +23,66 @@ public class CarServiceImpl implements CarService {
     private CarDao carDao;
 
     @Override
-    public Car findCarById(Long id) {
-        return carDao.findCarById(id);
+    public Car findCarById(Long id) throws DataAccessException {
+        if (id == null) throw new IllegalArgumentException("find by id argument is null");
+        try {
+            return carDao.findCarById(id);
+        } catch (Throwable throwable) {
+            throw new CustomDataAccessException("could not find by id car", throwable);
+        }
     }
 
     @Override
-    public Car findCarByDriver(Driver driver) {
-        return carDao.findCarByDriver(driver);
+    public Car findCarByDriver(Driver driver) throws DataAccessException {
+        if (driver == null) throw new IllegalArgumentException("find by driver argument is null");
+        try {
+            return carDao.findCarByDriver(driver);
+        } catch (Throwable throwable) {
+            throw new CustomDataAccessException("could not find by driver car", throwable);
+        }
     }
 
     @Override
-    public List<Car> listAllCars() {
-        return carDao.listAllCars();
+    public List<Car> listAllCars() throws DataAccessException {
+        try {
+            return carDao.listAllCars();
+        } catch (Throwable throwable) {
+            throw new CustomDataAccessException("could not list all cars", throwable);
+        }
     }
 
     @Override
-    public void createCar(Car car) {
-        carDao.addCar(car);
+    public void createCar(Car car) throws DataAccessException {
+        if (car == null) throw new IllegalArgumentException("car create argument is null");
+        try {
+            carDao.addCar(car);
+        } catch (Throwable throwable) {
+            throw new CustomDataAccessException("could not create car", throwable);
+        }
     }
 
     @Override
-    public void updateCar(Car car) {
-        carDao.updateCar(car);
+    public void updateCar(Car car) throws DataAccessException {
+        if (car == null) throw new IllegalArgumentException("car update argument is null");
+        try {
+            carDao.updateCar(car);
+        } catch (Throwable throwable) {
+            throw new CustomDataAccessException("could not update car", throwable);
+        }
     }
 
     @Override
-    public void deleteCar(Car car) {
-        carDao.deleteCar(car);
+    public void deleteCar(Car car) throws DataAccessException {
+        if (car == null) throw new IllegalArgumentException("car delete argument is null");
+        try {
+            carDao.deleteCar(car);
+        } catch (Throwable throwable) {
+            throw new CustomDataAccessException("could not delete car", throwable);
+        }
     }
 
     @Override
-    public List<Car> listCarsByComponentName(String componentName) {
+    public List<Car> listCarsByComponentName(String componentName) throws DataAccessException {
         List<Car> cars = listAllCars();
         List<Car> aero = cars.stream().filter(p -> p.getAerodynamics().getName().equals(componentName)).collect(Collectors.toList());
         List<Car> engine = cars.stream().filter(p -> p.getEngine().getName().equals(componentName)).collect(Collectors.toList());
